modularize tests
