+++ tests/test_convert.py
@@ -4,10 +4,10 @@ Tests for CytoTable.convert and related.
 
 # pylint: disable=no-member
 
-import itertools
+
 import pathlib
-from shutil import copy
-from typing import Any, Dict, List, Tuple, cast
+
+from typing import Any, Dict, List, cast
 
 import duckdb
 import parsl
@@ -17,486 +17,17 @@ from parsl.channels import LocalChannel
 from parsl.config import Config
 from parsl.executors import HighThroughputExecutor
 from parsl.providers import LocalProvider
-from pyarrow import csv, parquet
+from pyarrow import parquet
 from pycytominer.cyto_utils.cells import SingleCells
 
-from cytotable.apps.combine import (
-    _concat_join_sources,
-    _concat_source_group,
-    _get_join_chunks,
-    _infer_source_group_common_schema,
-    _join_source_chunk,
-)
-from cytotable.apps.modify import _prepend_column_name
-from cytotable.apps.path import _get_filepaths, _infer_path_datatype
-from cytotable.apps.workflow import _to_parquet
 from cytotable.convert import convert
-from cytotable.presets import config
+
 from cytotable.utils import (
     _column_sort,
     _duckdb_reader,
     _sqlite_mixed_type_query_to_parquet,
+    _default_parsl_config,
 )
-
-
-def test_config():
-    """
-    Tests config to ensure proper values
-    """
-
-    # check that we have relevant keys for each preset
-    for config_preset in config.values():
-        assert sorted(
-            [
-                "CONFIG_NAMES_COMPARTMENTS",
-                "CONFIG_NAMES_METADATA",
-                "CONFIG_IDENTIFYING_COLUMNS",
-                "CONFIG_CHUNK_SIZE",
-                "CONFIG_CHUNK_COLUMNS",
-                "CONFIG_JOINS",
-                "CONFIG_SOURCE_VERSION",
-            ]
-        ) == sorted(config_preset.keys())
-
-
-def test_get_filepaths(get_tempdir: str, data_dir_cellprofiler: str):
-    """
-    Tests _get_filepaths
-    """
-
-    # test that no sources raises an exception
-    empty_dir = pathlib.Path(f"{get_tempdir}/temp")
-    empty_dir.mkdir(parents=True, exist_ok=True)
-    with pytest.raises(Exception):
-        single_dir_result = _get_filepaths(
-            path=empty_dir,
-            targets=["image", "cells", "nuclei", "cytoplasm"],
-        ).result()
-
-    # check that single sqlite file is returned as desired
-    single_file_result = _get_filepaths(
-        path=pathlib.Path(
-            f"{data_dir_cellprofiler}/NF1_SchwannCell_data/all_cellprofiler.sqlite"
-        ),
-        targets=["cells"],
-    ).result()
-    assert len(set(single_file_result.keys())) == 1
-
-    # check that single csv file is returned as desired
-    single_file_result = _get_filepaths(
-        path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman/Cells.csv"),
-        targets=["cells"],
-    ).result()
-    assert len(set(single_file_result.keys())) == 1
-
-    single_dir_result = _get_filepaths(
-        path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman"),
-        targets=["cells"],
-    ).result()
-    # test that the single dir structure includes 1 unique key (for cells)
-    assert len(set(single_dir_result.keys())) == 1
-
-    single_dir_result = _get_filepaths(
-        path=pathlib.Path(f"{data_dir_cellprofiler}/ExampleHuman"),
-        targets=["image", "cells", "nuclei", "cytoplasm"],
-    ).result()
-    # test that the single dir structure includes 4 unique keys
-    assert len(set(single_dir_result.keys())) == 4
-
-
-def test_prepend_column_name(get_tempdir: str):
-    """
-    Tests _prepend_column_name
-    """
-
-    # example cytoplasm csv table run
-    prepend_testpath_1 = f"{get_tempdir}/prepend_testpath_1.parquet"
-    parquet.write_table(
-        table=pa.Table.from_pydict(
-            {
-                "ImageNumber": [1, 2, 3, 1, 2, 3],
-                "ObjectNumber": [1, 1, 1, 2, 2, 2],
-                "Parent_Cells": [1, 1, 1, 2, 2, 2],
-                "Parent_Nuclei": [1, 1, 1, 2, 2, 2],
-                "field1": ["foo", "bar", "baz", "foo", "bar", "baz"],
-                "field2": [True, False, True, True, False, True],
-            }
-        ),
-        where=prepend_testpath_1,
-    )
-    result = _prepend_column_name(
-        table_path=prepend_testpath_1,
-        source_group_name="Cytoplasm.csv",
-        identifying_columns=[
-            "ImageNumber",
-            "ObjectNumber",
-            "Parent_Cells",
-            "Parent_Nuclei",
-        ],
-        metadata=["image"],
-        compartments=["image", "cells", "nuclei", "cytoplasm"],
-    ).result()
-
-    # compare the results with what's expected for column names
-    assert parquet.read_table(source=result).column_names == [
-        "Metadata_ImageNumber",
-        "Metadata_ObjectNumber",
-        "Metadata_Cytoplasm_Parent_Cells",
-        "Metadata_Cytoplasm_Parent_Nuclei",
-        "Cytoplasm_field1",
-        "Cytoplasm_field2",
-    ]
-
-    # example cells sqlite table run
-    repend_testpath_2 = f"{get_tempdir}/prepend_testpath_2.parquet"
-    parquet.write_table(
-        table=pa.Table.from_pydict(
-            {
-                "ImageNumber": [1, 2, 3, 1, 2, 3],
-                "Cells_Number_Object_Number": [1, 1, 1, 2, 2, 2],
-                "Parent_OrigNuclei": [1, 1, 1, 2, 2, 2],
-                "field1": ["foo", "bar", "baz", "foo", "bar", "baz"],
-                "field2": [True, False, True, True, False, True],
-            }
-        ),
-        where=repend_testpath_2,
-    )
-    result = _prepend_column_name(
-        table_path=repend_testpath_2,
-        source_group_name="Per_Cells.sqlite",
-        identifying_columns=[
-            "ImageNumber",
-            "Parent_Cells",
-            "Parent_OrigNuclei",
-        ],
-        metadata=["image"],
-        compartments=["image", "cells", "nuclei", "cytoplasm"],
-    ).result()
-
-    # compare the results with what's expected for column names
-    assert parquet.read_table(source=result).column_names == [
-        "Metadata_ImageNumber",
-        "Cells_Number_Object_Number",
-        "Metadata_Cells_Parent_OrigNuclei",
-        "Cells_field1",
-        "Cells_field2",
-    ]
-
-
-def test_concat_source_group(
-    get_tempdir: str,
-    example_tables: Tuple[pa.Table, ...],
-    example_local_sources: Dict[str, List[Dict[str, Any]]],
-):
-    """
-    Tests _concat_source_group
-    """
-
-    _, _, _, table_nuclei_1, table_nuclei_2 = example_tables
-
-    # simulate concat
-    concat_table = pa.concat_tables([table_nuclei_1, table_nuclei_2])
-
-    result = _concat_source_group(
-        source_group_name="nuclei",
-        source_group=example_local_sources["nuclei.csv"],
-        dest_path=get_tempdir,
-        common_schema=table_nuclei_1.schema,
-    ).result()
-    assert len(result) == 1
-    assert parquet.read_schema(result[0]["table"][0]) == concat_table.schema
-    assert (
-        parquet.read_metadata(result[0]["table"][0]).num_rows,
-        parquet.read_metadata(result[0]["table"][0]).num_columns,
-    ) == concat_table.shape
-
-    # add a mismatching source to group
-    mismatching_table = pa.Table.from_pydict(
-        {
-            "color": pa.array(["blue", "red", "green", "orange"]),
-        }
-    )
-    pathlib.Path(f"{get_tempdir}/example/5").mkdir(parents=True, exist_ok=True)
-    csv.write_csv(mismatching_table, f"{get_tempdir}/example/5/nuclei.csv")
-    parquet.write_table(mismatching_table, f"{get_tempdir}/example/5.nuclei.parquet")
-    example_local_sources["nuclei.csv"].append(
-        {
-            "source_path": pathlib.Path(f"{get_tempdir}/example/5/nuclei.csv"),
-            "destination_path": pathlib.Path(f"{get_tempdir}/example/5.nuclei.parquet"),
-        }
-    )
-
-    with pytest.raises(Exception):
-        _concat_source_group(
-            source_group_name="nuclei",
-            source_group=example_local_sources["nuclei.csv"],
-            dest_path=get_tempdir,
-        ).result()
-
-
-def test_get_join_chunks(get_tempdir: str):
-    """
-    Tests _get_join_chunks
-    """
-
-    # form test path
-    test_path = f"{get_tempdir}/merge_chunks_test.parquet"
-
-    # write test data to file
-    parquet.write_table(
-        table=pa.Table.from_pydict(
-            {
-                "id1": [1, 2, 3, 1, 2, 3],
-                "id2": ["a", "a", "a", "b", "b", "b"],
-                "field1": ["foo", "bar", "baz", "foo", "bar", "baz"],
-                "field2": [True, False, True, True, False, True],
-            }
-        ),
-        where=test_path,
-    )
-
-    result = _get_join_chunks(
-        sources={"merge_chunks_test.parquet": [{"table": [test_path]}]},
-        metadata=["merge_chunks_test"],
-        chunk_columns=["id1", "id2"],
-        chunk_size=2,
-    ).result()
-
-    # test that we have 3 chunks of merge columns
-    assert len(result) == 3
-    # test that we have only the columns we specified
-    assert set(
-        itertools.chain(
-            *[list(chunk_item.keys()) for chunk in result for chunk_item in chunk]
-        )
-    ) == {"id1", "id2"}
-
-
-def test_join_source_chunk(get_tempdir: str):
-    """
-    Tests _get_join_chunks
-    """
-
-    # form test path a
-    test_path_a = f"{get_tempdir}/example_a_merged.parquet"
-    # form test path b
-    test_path_b = f"{get_tempdir}/example_b_merged.parquet"
-
-    # write test data to file
-    parquet.write_table(
-        table=pa.Table.from_pydict(
-            {
-                "id1": [1, 2, 3, 1, 2, 3],
-                "id2": ["a", "a", "a", "b", "b", "b"],
-                "field1": ["foo", "bar", "baz", "foo", "bar", "baz"],
-            }
-        ),
-        where=test_path_a,
-    )
-    # write test data to file
-    parquet.write_table(
-        table=pa.Table.from_pydict(
-            {
-                "id1": [1, 2, 3, 1, 2, 3],
-                "id2": ["a", "a", "a", "b", "b", "b"],
-                "field2": [True, False, True, True, False, True],
-            }
-        ),
-        where=test_path_b,
-    )
-
-    result = _join_source_chunk(
-        sources={
-            "example_a": [{"table": [test_path_a]}],
-            "example_b": [{"table": [test_path_b]}],
-        },
-        dest_path=f"{get_tempdir}/destination.parquet",
-        joins=f"""
-            SELECT *
-            FROM read_parquet('{get_tempdir}/example_a_merged.parquet') as example_a
-            JOIN read_parquet('{get_tempdir}/example_b_merged.parquet') as example_b ON
-                example_b.id1 = example_a.id1
-                AND example_b.id2 = example_a.id2
-        """,
-        join_group=[{"id1": 1, "id2": "a"}, {"id1": 2, "id2": "a"}],
-        drop_null=True,
-    ).result()
-
-    assert isinstance(result, str)
-    result_table = parquet.read_table(source=result)
-    assert result_table.equals(
-        other=pa.Table.from_pydict(
-            {
-                "id1": [1, 2],
-                "id2": ["a", "a"],
-                "field1": ["foo", "bar"],
-                "field2": [True, False],
-            },
-            # use schema from result as a reference for col order
-            schema=result_table.schema,
-        )
-    )
-
-
-def test_concat_join_sources(get_tempdir: str):
-    """
-    Tests _concat_join_sources
-    """
-
-    # create a test dir
-    pathlib.Path(f"{get_tempdir}/concat_join/").mkdir(exist_ok=True)
-
-    # form test paths
-    test_path_a = f"{get_tempdir}/concat_join/join_chunks_test_a.parquet"
-    test_path_b = f"{get_tempdir}/concat_join/join_chunks_test_b.parquet"
-    test_path_a_join_chunk = f"{get_tempdir}/join_chunks_test_a.parquet"
-    test_path_b_join_chunk = f"{get_tempdir}/join_chunks_test_b.parquet"
-
-    # form test data
-    test_table_a = pa.Table.from_pydict(
-        {
-            "id1": [
-                1,
-                2,
-                3,
-            ],
-            "id2": [
-                "a",
-                "a",
-                "a",
-            ],
-            "field1": [
-                "foo",
-                "bar",
-                "baz",
-            ],
-            "field2": [
-                True,
-                False,
-                True,
-            ],
-        }
-    )
-    test_table_b = pa.Table.from_pydict(
-        {
-            "id1": [1, 2, 3],
-            "id2": ["b", "b", "b"],
-            "field1": ["foo", "bar", "baz"],
-            "field2": [True, False, True],
-        }
-    )
-
-    # write test data to file
-    parquet.write_table(
-        table=test_table_a,
-        where=test_path_a,
-    )
-    parquet.write_table(
-        table=test_table_b,
-        where=test_path_b,
-    )
-
-    # copy the files for testing purposes
-    copy(test_path_a, test_path_a_join_chunk)
-    copy(test_path_b, test_path_b_join_chunk)
-
-    pathlib.Path(f"{get_tempdir}/test_concat_join_sources").mkdir(
-        parents=True, exist_ok=True
-    )
-
-    result = _concat_join_sources(
-        dest_path=f"{get_tempdir}/test_concat_join_sources/example_concat_join.parquet",
-        join_sources=[test_path_a_join_chunk, test_path_b_join_chunk],
-        sources={
-            "join_chunks_test_a.parquet": [{"table": [test_path_a]}],
-            "join_chunks_test_b.parquet": [{"table": [test_path_b]}],
-        },
-    ).result()
-
-    # ensure the concatted result is what we expect
-    assert parquet.read_table(source=result).equals(
-        pa.concat_tables(tables=[test_table_a, test_table_b])
-    )
-
-    # ensure the test paths provided via sources were removed (unlinked)
-    assert (pathlib.Path(test_path_a).exists() is False) and (
-        pathlib.Path(test_path_a).exists() is False
-    )
-
-
-def test_infer_path_datatype():
-    """
-    Tests _infer_path_datatype
-    """
-
-    data = {
-        "sample_1.csv": [{"source_path": "stub"}],
-        "sample_2.CSV": [{"source_path": "stub"}],
-    }
-    assert _infer_path_datatype(sources=data).result() == "csv"
-    with pytest.raises(Exception):
-        _infer_path_datatype(sources=data, source_datatype="parquet").result()
-
-    data["sample_3.parquet"] = [{"source_path": "stub"}]
-    assert (
-        _infer_path_datatype(sources=data, source_datatype="parquet").result()
-        == "parquet"
-    )
-    with pytest.raises(Exception):
-        _infer_path_datatype(sources=data).result()
-
-
-def test_to_parquet(
-    get_tempdir: str, example_local_sources: Dict[str, List[Dict[str, Any]]]
-):
-    """
-    Tests _to_parquet
-    """
-
-    flattened_example_sources = list(
-        itertools.chain(*list(example_local_sources.values()))
-    )
-
-    # note: we cast here for mypy linting (dict and str treatment differ)
-    result: Dict[str, List[Dict[str, Any]]] = cast(
-        dict,
-        _to_parquet(
-            source_path=str(
-                example_local_sources["image.csv"][0]["source_path"].parent
-            ),
-            dest_path=get_tempdir,
-            source_datatype=None,
-            compartments=["cytoplasm", "cells", "nuclei"],
-            metadata=["image"],
-            identifying_columns=["imagenumber"],
-            concat=False,
-            join=False,
-            joins=None,
-            chunk_columns=None,
-            chunk_size=4,
-            infer_common_schema=False,
-            drop_null=True,
-        ).result(),
-    )
-
-    flattened_results = list(itertools.chain(*list(result.values())))
-    for i, flattened_result in enumerate(flattened_results):
-        parquet_result = parquet.ParquetDataset(
-            path_or_paths=flattened_result["table"]
-        ).read()
-        csv_source = (
-            _duckdb_reader()
-            .execute(
-                f"""
-                select * from
-                read_csv_auto('{str(flattened_example_sources[i]["source_path"])}',
-                ignore_errors=TRUE)
-                """
-            )
-            .arrow()
-        )
-        assert parquet_result.schema.equals(csv_source.schema)
-        assert parquet_result.shape == csv_source.shape
 
 
 def test_convert_s3_path_csv(
@@ -600,22 +131,6 @@ def test_convert_s3_path_sqlite(
 
     assert local_cytotable_table.equals(s3_cytotable_table)
     assert local_cytotable_table.equals(s3_cytotable_table_nested)
-
-
-def test_infer_source_group_common_schema(
-    example_local_sources: Dict[str, List[Dict[str, Any]]],
-    example_tables: Tuple[pa.Table, ...],
-):
-    """
-    Tests _infer_source_group_common_schema
-    """
-    _, _, _, table_nuclei_1, _ = example_tables
-
-    result = _infer_source_group_common_schema(
-        source_group=example_local_sources["nuclei.csv"],
-    ).result()
-
-    assert table_nuclei_1.schema.equals(pa.schema(result))
 
 
 def test_convert_cytominerdatabase_csv(
@@ -939,3 +454,4 @@ def test_convert_hte_cellprofiler_csv(
 
     # clean up the parsl config for other tests
     parsl.clear()
+    parsl.load(_default_parsl_config())
